Update style.css
added opacity to the video clip I added so it wont be too bright.  Again this is optional and if this would disrupt the format of the webpage, it is ok not to add.  I wasn't sure about the z-index function if I did it correctly.  I wanted to place the video clip in the back as a background behind everything else.